# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Skew AI — Voice-of-Customer Intelligence Platform
  description: 'Skew AI (skewai): domain-agnostic voice-of-customer intelligence.
    Plug in a Domain Pack and get an AI voice agent, a swarm of triage/RCA agents,
    and the Qubot v2 auditor.


    ## Auth

    When `FRONTLINE_API_KEY` is set, send `X-API-Key` or `Authorization: Bearer`.


    ## Versioning

    Current API version **1** (header `API-Version`). Paths under `/api/...` are v1
    without a URL prefix. Breaking changes will introduce `/api/v2/...`.


    ## Errors

    Errors use RFC 7807-shaped JSON (`type`, `title`, `status`, `detail`) while keeping
    FastAPI''s `detail` string for existing clients.


    ## Pagination

    Collection endpoints accept `limit`, `offset`, and optional `cursor`, and return
    a `pagination` object (`has_more`, `next_offset`, `next_cursor`).'
  version: 2.0.0
paths:
...
  /api/frontline/analytics/forecast:
    get:
      tags:
      - frontline-platform
      summary: Analytics Forecast
      operationId: analytics_forecast_api_frontline_analytics_forecast_get
      parameters:
      - name: pack_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Pack Id
      - name: window_days
        in: query
        required: false
        schema:
          type: integer
          default: 28
          title: Window Days
      - name: threshold
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Threshold
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Analytics Forecast Api Frontline Analytics Forecast
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/frontline/analytics/cross-pack:
    get:
      tags:
      - frontline-platform
      summary: Analytics Cross Pack
      operationId: analytics_cross_pack_api_frontline_analytics_cross_pack_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Analytics Cross Pack Api Frontline Analytics Cross
                  Pack Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/frontline/analytics/hotspots:
    get:
      tags:
      - frontline-platform
      summary: Analytics Hotspots
      operationId: analytics_hotspots_api_frontline_analytics_hotspots_get
      parameters:
      - name: pack_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Pack Id
      - name: window_days
        in: query
        required: false
        schema:
          type: integer
          default: 90
          title: Window Days
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Analytics Hotspots Api Frontline Analytics Hotspots
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/frontline/analytics/severity-drift:
    get:
      tags:
      - frontline-platform
      summary: Analytics Severity Drift
      operationId: analytics_severity_drift_api_frontline_analytics_severity_drift_get
      parameters:
      - name: pack_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Pack Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Analytics Severity Drift Api Frontline Analytics Severity
                  Drift Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/frontline/analytics/cohorts:
    get:
      tags:
      - frontline-platform
      summary: Analytics Cohorts
      operationId: analytics_cohorts_api_frontline_analytics_cohorts_get
      parameters:
      - name: pack_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Pack Id
      - name: group_field
        in: query
        required: false
        schema:
          type: string
          default: entity_3
          title: Group Field
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Analytics Cohorts Api Frontline Analytics Cohorts
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/frontline/analytics/regulator-watch:
    get:
      tags:
      - frontline-platform
      summary: Analytics Regulator
      operationId: analytics_regulator_api_frontline_analytics_regulator_watch_get
      parameters:
      - name: pack_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Pack Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Analytics Regulator Api Frontline Analytics Regulator
                  Watch Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/frontline/analytics/financial-impact:
    get:
      tags:
      - frontline-platform
      summary: Analytics Financial
      operationId: analytics_financial_api_frontline_analytics_financial_impact_get
      parameters:
      - name: pack_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Pack Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Analytics Financial Api Frontline Analytics Financial
                  Impact Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/frontline/analytics/fairness:
    get:
      tags:
      - frontline-platform
      summary: Analytics Fairness
      operationId: analytics_fairness_api_frontline_analytics_fairness_get
      parameters:
      - name: pack_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Pack Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Analytics Fairness Api Frontline Analytics Fairness
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/frontline/booking/offer:
    post:
      tags:
      - frontline-platform
      summary: Booking Offer
      operationId: booking_offer_api_frontline_booking_offer_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              title: Body
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Booking Offer Api Frontline Booking Offer Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/frontline/booking/book:
    post:
      tags:
      - frontline-platform
      summary: Booking Book
      operationId: booking_book_api_frontline_booking_book_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              title: Body
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Booking Book Api Frontline Booking Book Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/frontline/callbacks:
    post:
      tags:
      - frontline-platform
      summary: Callbacks Create
      operationId: callbacks_create_api_frontline_callbacks_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              title: Body
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Callbacks Create Api Frontline Callbacks Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - frontline-platform
      summary: Callbacks List
      operationId: callbacks_list_api_frontline_callbacks_get
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          default: scheduled
          title: Status
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Callbacks List Api Frontline Callbacks Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/frontline/marketplace/packs:
    get:
      tags:
      - frontline-platform
      summary: Marketplace List
      operationId: marketplace_list_api_frontline_marketplace_packs_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Marketplace List Api Frontline Marketplace Packs Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...